Slet biledfiler

+++ ZealandZooAPP/Pages/Gallery/Gallery.cshtml.cs
@@ -8,11 +8,13 @@ namespace ZealandZooAPP.Pages.Gallery
     public class GalleryModel : PageModel
     {
         private readonly ImageRepoService _imageRepoService;
+        private readonly IFileService _localFileService;
         public List<BaseModel> Images { get; set; }
 
-        public GalleryModel(ImageRepoService imageRepoService)
+        public GalleryModel(ImageRepoService imageRepoService, IFileService localFileService )
         {
             _imageRepoService = imageRepoService;
+            _localFileService = localFileService;
         }
 
         public void OnGet()
@@ -22,7 +24,12 @@ namespace ZealandZooAPP.Pages.Gallery
 
         public void OnPostDeleteImage(int id)
         {
-            _imageRepoService.Delete(id);
+            EventImage imageToBeDeleted = (EventImage)_imageRepoService.GetById(id);
+
+            if (_localFileService.Delete(imageToBeDeleted.Name))
+            {
+                _imageRepoService.Delete(id);
+            }
 
             Images = _imageRepoService.GetAll();
